Don't track YGNode pointers on the Haskell side.
Before, we kept track of the child pointers for each node in the tree structure
of the 'Layout' datatype. This was causing problems with bookkeeping due to
the somewhat complicated nature of the Haskell runtime. This change allows us to
keep that bookkeeping within the Yoga library itself, and traverse the tree
structure as necessary during rendering. Now that Yoga has a node cloning
function, it becomes much easier to assign children to parents without
disrupting the immutability that Haskell expects from its pure functions. This
change shouldn't affect any of the APIs that depend on this library. The tree
structure that is tracked on the Haskell side still matches that of the Yoga
library internally, but we only keep Haskell values on this side of the fence.