Refactor the high-level layout definition.
Instead of storing a tree directly, store a computation that produces a
tree. This has two main benefits:

1. We can structure our layout update calls in terms of adding a
computation onto the end of the layout generation. This means that we
can avoid calling `unsafePerformIO` until we actually need it, which
is during rendering, at which point we make it explicit.

2. The layout computation actually only happens once, the first time
the layout is rendered. Each subsequent time, the layout is cached and
no re-layout-ing is required.

This also gets rid of a lot of the weird type signatures that we were
using to try and keep the semantics of our layout definition memory safe
and (obviously) removes the need for `unsafePerformIO`.

Fixes #4.